pyproject.toml

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "", email = ""},
]
dependencies = [
    "flask>=2.3.2",
    "youtube-dl>=2021.12.17",
    "openai-whisper @ git+https://github.com/openai/whisper.git",
    "yt-dlp>=2023.7.6",
    "openai>=0.27.8",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}
